% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_bootstrap_see}
\alias{calculate_bootstrap_see}
\title{Calculate Bootstrap Standard Error of Equating}
\usage{
calculate_bootstrap_see(boot_matrix)
}
\arguments{
\item{boot_matrix}{Matrix of bootstrap equated scores (nsim x n_scores)}
}
\value{
Named vector of SEE values for each score
}
\description{
Computes the standard error of equating from bootstrap samples
using the standard deviation of the bootstrap distribution.
}
\keyword{internal}
