% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_direct.R
\name{calculate_expected_score}
\alias{calculate_expected_score}
\title{Calculate expected score given theta}
\usage{
calculate_expected_score(theta, gamma, score_min, score_max)
}
\description{
Calculate expected score given theta
}
\keyword{internal}
