% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{calculate_statistics}
\alias{calculate_statistics}
\title{Calculate fit statistics (CALCULATE_STATISTICS from Pascal)}
\usage{
calculate_statistics(object)
}
\arguments{
\item{object}{A leunbach_ipf object}
}
\value{
A list with fit statistics
}
\description{
Computes the likelihood ratio test statistic, Goodman-Kruskal Gamma test,
and degrees of freedom following the DIGRAM implementation.
}
\keyword{internal}
