% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_direct.R
\name{calculate_true_score}
\alias{calculate_true_score}
\title{Calculate expected (true) score on target test given theta}
\usage{
calculate_true_score(theta, gamma, score_min, score_max)
}
\description{
Calculate expected (true) score on target test given theta
}
\keyword{internal}
