% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_error_frequencies}
\alias{compute_error_frequencies}
\title{Compute frequency of bootstrap equating errors}
\usage{
compute_error_frequencies(boot_rounded, observed_rounded, scores)
}
\arguments{
\item{boot_rounded}{Matrix of bootstrap rounded scores (nsim x n_scores)}

\item{observed_rounded}{Vector of observed rounded scores}

\item{scores}{Vector of source scores}
}
\value{
Matrix of error frequencies (n_scores x 5) for errors -2, -1, 0, +1, +2
}
\description{
Computes the frequency distribution of equating errors (differences between
bootstrap equated scores and observed equated scores) for each source score.
}
\keyword{internal}
