% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{compute_indirect_equating}
\alias{compute_indirect_equating}
\title{Compute indirect equating table using interpolation}
\usage{
compute_indirect_equating(eq_ab, eq_bc)
}
\description{
Compute indirect equating table using interpolation
}
\keyword{internal}
