% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_direct.R
\name{estimate_theta_bisection}
\alias{estimate_theta_bisection}
\title{Estimate theta using bisection method}
\usage{
estimate_theta_bisection(
  score,
  gamma,
  score_min,
  score_max,
  tol = 1e-06,
  max_iter = 100
)
}
\description{
Estimate theta using bisection method
}
\keyword{internal}
