% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_direct.R
\name{estimate_theta_newton}
\alias{estimate_theta_newton}
\title{Estimate theta using Newton-Raphson with bisection fallback}
\usage{
estimate_theta_newton(
  score,
  gamma,
  score_min,
  score_max,
  tol = 1e-07,
  max_iter = 500
)
}
\description{
Estimate theta using Newton-Raphson with bisection fallback
}
\keyword{internal}
