% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_direct.R
\name{estimate_theta_optimize}
\alias{estimate_theta_optimize}
\title{Estimate theta using stats::optimize() with Brent's method}
\usage{
estimate_theta_optimize(score, gamma, score_min, score_max, tol = 1e-10)
}
\description{
Estimate theta using stats::optimize() with Brent's method
}
\keyword{internal}
