% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_direct.R
\name{expected_score_and_deriv}
\alias{expected_score_and_deriv}
\title{Calculate expected score and its derivative with respect to theta}
\usage{
expected_score_and_deriv(theta, gamma, score_min, score_max)
}
\description{
Calculate expected score and its derivative with respect to theta
}
\keyword{internal}
