% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{find_nearest_equated}
\alias{find_nearest_equated}
\title{Find nearest valid equated score}
\usage{
find_nearest_equated(score, lookup, direction = c("down", "up"))
}
\description{
Find nearest valid equated score
}
\keyword{internal}
