% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leunbachR-package.R
\docType{package}
\name{leunbachR-package}
\alias{leunbachR}
\alias{leunbachR-package}
\title{leunbachR: Leunbach Test Equating}
\description{
Implements the Leunbach test equating method, following the DIGRAM software
written by Svend Kreiner. Both direct and indirect equating are available,
with parametric bootstrap standard errors and diagnostic statistics
including the Goodman-Kruskal gamma test and orbit analysis for person fit.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pgmj/leunbachR}
  \item \url{https://pgmj.github.io/leunbachR/}
  \item Report bugs at \url{https://github.com/pgmj/leunbachR/issues}
}

}
\author{
\strong{Maintainer}: Magnus Johansson \email{pgmj@pm.me} (\href{https://orcid.org/0000-0003-1669-592X}{ORCID})

}
\keyword{internal}
