% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{print.leunbach_indirect}
\alias{print.leunbach_indirect}
\title{Print method for leunbach_indirect objects}
\usage{
\method{print}{leunbach_indirect}(x, ...)
}
\arguments{
\item{x}{A \code{leunbach_indirect} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for leunbach_indirect objects
}
