% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{print.leunbach_indirect_bootstrap}
\alias{print.leunbach_indirect_bootstrap}
\title{Print method for leunbach_indirect_bootstrap objects}
\usage{
\method{print}{leunbach_indirect_bootstrap}(x, ...)
}
\arguments{
\item{x}{A \code{leunbach_indirect_bootstrap} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for leunbach_indirect_bootstrap objects
}
