% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{print.leunbach_ipf}
\alias{print.leunbach_ipf}
\title{Print method for leunbach_ipf objects}
\usage{
\method{print}{leunbach_ipf}(x, ...)
}
\arguments{
\item{x}{A \code{leunbach_ipf} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for leunbach_ipf objects
}
