% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_direct.R
\name{run_bootstrap_parallel}
\alias{run_bootstrap_parallel}
\title{Run bootstrap samples in parallel using mirai}
\usage{
run_bootstrap_parallel(
  nsim,
  boot_seeds,
  boot_data_list,
  n_cores,
  verbose = FALSE
)
}
\description{
Run bootstrap samples in parallel using mirai
}
\keyword{internal}
