% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_direct.R
\name{run_bootstrap_sequential}
\alias{run_bootstrap_sequential}
\title{Run bootstrap samples sequentially}
\usage{
run_bootstrap_sequential(nsim, boot_seeds, boot_data_list, verbose = FALSE)
}
\description{
Run bootstrap samples sequentially
}
\keyword{internal}
