% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{run_indirect_bootstrap_parallel}
\alias{run_indirect_bootstrap_parallel}
\title{Run indirect bootstrap in parallel}
\usage{
run_indirect_bootstrap_parallel(
  nsim,
  boot_seeds,
  boot_data_list,
  n_cores,
  verbose = FALSE
)
}
\description{
Run indirect bootstrap in parallel
}
\keyword{internal}
