% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{run_indirect_bootstrap_sequential}
\alias{run_indirect_bootstrap_sequential}
\title{Run indirect bootstrap sequentially}
\usage{
run_indirect_bootstrap_sequential(
  nsim,
  boot_seeds,
  boot_data_list,
  verbose = FALSE
)
}
\description{
Run indirect bootstrap sequentially
}
\keyword{internal}
