% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{summary.leunbach_indirect_bootstrap}
\alias{summary.leunbach_indirect_bootstrap}
\title{Summary method for leunbach_indirect_bootstrap objects}
\usage{
\method{summary}{leunbach_indirect_bootstrap}(object, ...)
}
\arguments{
\item{object}{A \code{leunbach_indirect_bootstrap} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for leunbach_indirect_bootstrap objects
}
