% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{summary.leunbach_ipf}
\alias{summary.leunbach_ipf}
\title{Summary method for leunbach_ipf objects}
\usage{
\method{summary}{leunbach_ipf}(object, ...)
}
\arguments{
\item{object}{A \code{leunbach_ipf} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns a list of fit statistics.
}
\description{
Summary method for leunbach_ipf objects
}
