% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbits.R
\name{summary.leunbach_orbits}
\alias{summary.leunbach_orbits}
\title{Summary method for leunbach_orbits objects}
\usage{
\method{summary}{leunbach_orbits}(object, ...)
}
\arguments{
\item{object}{A \code{leunbach_orbits} object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for leunbach_orbits objects
}
