% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{diagnose_equating}
\alias{diagnose_equating}
\title{Diagnostic function for equating}
\usage{
diagnose_equating(
  fit,
  direction = c("1to2", "2to1"),
  method = c("optimize", "newton")
)
}
\arguments{
\item{fit}{A leunbach_ipf object}

\item{direction}{Direction of equating:  "1to2" or "2to1"}

\item{method}{Optimization method: "optimize" or "newton"}
}
\description{
Diagnoses potential issues with person parameter estimation by displaying
theta values and expected scores for each source score.
}
