% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{get_indirect_equating_table}
\alias{get_indirect_equating_table}
\title{Get indirect equating table}
\usage{
get_indirect_equating_table(boot)
}
\arguments{
\item{boot}{A leunbach_indirect_bootstrap object}
}
\value{
A data frame with equating results and CIs
}
\description{
Get indirect equating table
}
