% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbits.R
\name{get_orbit}
\alias{get_orbit}
\title{Get orbit distribution for a specific total score}
\usage{
get_orbit(orbits, total_score)
}
\arguments{
\item{orbits}{A leunbach_orbits object}

\item{total_score}{The total score to examine}
}
\value{
A data frame with the orbit distribution
}
\description{
Get orbit distribution for a specific total score
}
