% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{plot.leunbach_indirect_bootstrap}
\alias{plot.leunbach_indirect_bootstrap}
\title{Plot method for leunbach_indirect_bootstrap objects}
\usage{
\method{plot}{leunbach_indirect_bootstrap}(x, type = c("equating", "see"), ...)
}
\description{
Plot method for leunbach_indirect_bootstrap objects
}
