% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipf.R
\name{plot.leunbach_ipf}
\alias{plot.leunbach_ipf}
\title{Plot method for leunbach_ipf objects}
\usage{
\method{plot}{leunbach_ipf}(x, type = c("parameters", "residuals", "observed", "sigma"), ...)
}
\description{
Plot method for leunbach_ipf objects
}
