% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbits.R
\name{plot.leunbach_orbits}
\alias{plot.leunbach_orbits}
\title{Plot method for leunbach_orbits objects}
\usage{
\method{plot}{leunbach_orbits}(x, type = c("orbits", "cumulative", "significant"), ...)
}
\description{
Plot method for leunbach_orbits objects
}
