% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_direct.R
\name{print.leunbach_bootstrap}
\alias{print.leunbach_bootstrap}
\title{Print method for leunbach_bootstrap objects}
\usage{
\method{print}{leunbach_bootstrap}(x, ...)
}
\description{
Print method for leunbach_bootstrap objects
}
