% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_direct.R
\name{print.leunbach_equating}
\alias{print.leunbach_equating}
\title{Print method for leunbach_equating objects}
\usage{
\method{print}{leunbach_equating}(x, ...)
}
\description{
Print method for leunbach_equating objects
}
