% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equate_indirect.R
\name{print.leunbach_indirect_bootstrap}
\alias{print.leunbach_indirect_bootstrap}
\title{Print method for leunbach_indirect_bootstrap objects}
\usage{
\method{print}{leunbach_indirect_bootstrap}(x, ...)
}
\description{
Print method for leunbach_indirect_bootstrap objects
}
