% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbits.R
\name{print.leunbach_orbits}
\alias{print.leunbach_orbits}
\title{Print method for leunbach_orbits objects}
\usage{
\method{print}{leunbach_orbits}(x, ...)
}
\description{
Print method for leunbach_orbits objects
}
